Add additional nlp tools and split into separate files
